user management & ui adjustment

+++ src/main/java/org/dadeco/cu996/api/model/Role.java
@@ -1,16 +1,9 @@
 package org.dadeco.cu996.api.model;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import lombok.*;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+import javax.persistence.*;
+import java.util.Set;
 
 @Entity
 @Setter
@@ -22,7 +15,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column
-    private Long id;
+    private Integer id;
 
     @Column(name = "name")
     @NonNull
